+++ db/migrate/20200319205106_add_symphony_id_to_location.rb
@@ -1,0 +1,6 @@
+class AddSymphonyIdToLocation < ActiveRecord::Migration
+  def change
+    # the Symphony id is used by SolrMarc to map from 596a to Institution
+    add_column :locations, :symphony_id, :integer, default: nil
+  end
+end