+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and releases in Discovery project adheres to [Semantic Versioning](http://semver
 ### Added
 - create backup tables for library and location [#1843](https://github.com/ualbertalib/discovery/issues/1843)
 - add symphony_nightly generator [#1843](https://github.com/ualbertalib/discovery/issues/1843)
+- add generator to prep files for SolrMarc before ingest [#1939](https://github.com/ualbertalib/discovery/issues/1939)
 
 ### Fixed
 - N+1 query memoized for location, status, item type and circulation rules [PR#1937](https://github.com/ualbertalib/discovery/pull/1937)